requirements.txt
exa_py==1.0.12
python-dotenv==1.0.1
streamlit==1.29.0